evaluation/backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
httpx>=0.28.0
pydantic>=2.10.0
openai>=2.21.0
python-dotenv>=1.2.1
sse-starlette>=2.2.0
pyyaml>=6.0
